Create standard client-side logger. - Use it for both the main client & the status page.
